Remove libipld dependency
This is a followup of https://github.com/blacksky-algorithms/rsky/pull/110.
`libipld` is still part of the dependency due to `rs-car-sync`. The most
recent version 0.5 switched to `ipld-core`. With this upgrade the whole
tree is now `libipld` free.

This also leads to a newer `cid` and `multihash` version, hence some code
changes are required.